@@ -1,10 +1,12 @@
 {
     "name": "im-parsa.github.io",
     "version": "2.0.0",
-    "homepage": ".",
+    "homepage": "https://parsa-firoozi.ir/",
     "scripts": {
         "start": "react-scripts start",
         "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject",
         "watch:sass": "sass src/Assets/Sass/Main.scss src/Assets/Sass/Stylesheets/Main.css -w"
     },
     "eslintConfig": {
@@ -26,24 +28,26 @@
         ]
     },
     "dependencies": {
-        "@types/react": "^17.0.34",
+        "@testing-library/jest-dom": "^5.16.1",
+        "@testing-library/react": "^12.1.2",
+        "@testing-library/user-event": "^13.5.0",
+        "@types/jest": "^27.0.3",
+        "@types/node": "^16.11.15",
+        "@types/react": "^17.0.37",
         "@types/react-dom": "^17.0.11",
         "@types/react-helmet": "^6.1.4",
-        "@types/react-router-dom": "^5.3.0",
         "axios": "^0.24.0",
         "html-react-parser": "^1.4.4",
         "moment": "^2.29.1",
+        "react": "^17.0.2",
         "react-countup": "^6.1.0",
         "react-dom": "^17.0.2",
         "react-helmet": "^6.1.0",
-        "react-markdown": "^7.1.1",
-        "react-router-dom": "^5.3.0",
-        "react-scripts": "^4.0.3",
+        "react-router-dom": "^6.2.1",
+        "react-scripts": "5.0.0",
         "sass": "^1.43.4",
-        "typescript": "^4.4.3",
+        "styled-components": "^5.3.3",
+        "typescript": "^4.5.4",
         "web-vitals": "^2.1.2"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.16.0"
     }
 }
